Add documentation about modifications to 'nucypher status events'. Minor refactor.
